requirements.txt
Django==4.1.3
djangorestframework==3.14.0
drf-yasg==1.21.4
django-filter==23.3